% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adecco-data.R
\docType{data}
\name{adecco}
\alias{adecco}
\title{Adecco Swiss Job Market Index Data}
\format{An object of class \code{swissdata} of length 3.}
\source{
\href{http://www.stellenmarktmonitor.uzh.ch/en/indices/asjmi.html}{University of Zürich}
}
\usage{
data(adecco)
}
\description{
Adecco Swiss job market index dataset in \code{swissdata} form
taken from the University of Zürich.
}
\examples{
data(adecco)
adecco$data
adecco$meta$hierarchy
adecco$meta$labels
}
\keyword{datasets}
