% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_kd.R
\name{calculate_KD_range}
\alias{calculate_KD_range}
\title{Calculate KD Range for Plotting}
\usage{
calculate_KD_range(data, n_lines)
}
\arguments{
\item{data}{A data frame containing k_off and k_on columns}

\item{n_lines}{Integer specifying the number of KD lines to generate}
}
\value{
A numeric vector of KD values
}
\description{
Helper function to calculate an appropriate range of KD values for plotting,
based on the input kinetic data.
}
\keyword{internal}
