% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_plasmid.R
\name{calculate_plasmid}
\alias{calculate_plasmid}
\title{Calculate Plasmid Amounts for Transfection}
\usage{
calculate_plasmid(
  plasmid_hc,
  conc_hc,
  transfection_volume,
  plasmid_lc = NA,
  conc_lc = NA,
  plasmid_vol = 1
)
}
\arguments{
\item{plasmid_hc}{A character vector of heavy chain plasmid names.}

\item{conc_hc}{A numeric vector of heavy chain plasmid concentrations (in µg/µL), matching the order of \code{plasmid_hc}.}

\item{transfection_volume}{A numeric scalar giving the total transfection volume per plasmid pair (in mL).}

\item{plasmid_lc}{(Optional) A character vector of light chain plasmid names, matching the length of \code{plasmid_hc} or shorter.
Missing entries are interpreted as no LC for those HC plasmids.}

\item{conc_lc}{(Optional) A numeric vector of light chain plasmid concentrations (in µg/µL), matching the order of \code{plasmid_lc}.
Must be the same length as \code{plasmid_lc}.}

\item{plasmid_vol}{A numeric scalar of total plasmid (µg) DNA that will be transfected per mL of culture.}
}
\value{
A tibble with columns:
\describe{
\item{pair_id}{An integer ID indicating each plasmid pair (or HC-only entry).}
\item{plasmid}{The name of the plasmid (HC or LC).}
\item{concentration}{The concentration of the plasmid (in µg/µL).}
\item{amount}{The calculated amount of plasmid to use (rounded to 0 decimal places, in µL).}
}
}
\description{
Computes the amount of plasmid (in µg) required for transfection based on DNA concentration and total transfection volume.
Supports both heavy chain (HC) only and HC-light chain (LC) plasmid pairs.
}
\examples{
calculate_plasmid(
  plasmid_hc = c("BACE1_HC", "BE10", "BF2", "BF8"),
  conc_hc = c(1.444, 1.822, 1.739, 1.703),
  plasmid_lc = c("BACE1_LC"),
  conc_lc = c(1.816),
  transfection_volume = 200,
  plasmid_vol = 1.0
)

}
