% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_plate.R
\name{create_plate}
\alias{create_plate}
\title{Creates a 96 well ELISA plate template}
\usage{
create_plate(
  plate_type = 96,
  n_plates = 21,
  plate_names = NULL,
  file_type = c("xlsx", "csv"),
  file_name = NULL
)
}
\arguments{
\item{plate_type}{An integer that defines the type of microwell plate (default
is 96).}

\item{n_plates}{An integer that defines the total number of plates (default
is 19).}

\item{plate_names}{A character vector that defines name of each plate.}

\item{file_type}{A character string that defines type of file (default is
xlsx)}

\item{file_name}{A character string that defines name of file}
}
\value{
A xlsx or csv file (default is xlsx)
}
\description{
Creates a 96 well ELISA plate template
}
\examples{
\dontrun{
create_plate()
}
}
