% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.R
\name{dataset_create}
\alias{dataset_create}
\title{Create Swissdata Objects}
\usage{
dataset_create(data, meta, dataset_id, test = TRUE)
}
\arguments{
\item{data}{data, or URL}

\item{meta}{meta, or URL}

\item{test}{should \code{dataset_validate()} be run on the newly-read object (default = TRUE)}
}
\value{
An object of class swissdata
}
\description{
Creator funtion for swissdata objects. All  swissdata objects should be
created by this function.
}
\examples{

dataset_create(adecco$data, adecco$meta, dataset_id = "ch_adecco_sjmi")
}
