% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_operations.R
\name{dataset_merge}
\alias{dataset_merge}
\title{Swissdata Dataset Operations}
\usage{
dataset_merge(x, y)
}
\arguments{
\item{x}{original swissdata object}

\item{y}{additional swissdata object}
}
\value{
a combined swissdata object
}
\description{
A set of functions for manipulating swissdata datasets.
}
\details{
\code{dataset_merge} combines the dimension levels from two swissdata datasets.

\code{dataset_merge} adds dimension levels that are present only in \code{y} to \code{x}.
As a result in the case where the same dimensions or levels are present
in both of the objects - the one from \code{x} is selected. \code{test_swissdata()}
function is executed before returning the output in order to make sure the
combination of the two objects was successful.
}
\examples{
# merge two distinct datasets
z1 <- adecco
str(z1)
z2 <- level_rename(z1, "idx_type", "ins", "new")
str(z2)
z <- dataset_merge(z1, z2)
str(z)

}
\author{
Christoph Sax
}
