% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dim_add.R
\name{dim_drop}
\alias{dim_drop}
\title{Drop Existing Dimension}
\usage{
dim_drop(x, dim = "trans", level = "ind")
}
\arguments{
\item{x}{swissdata object}

\item{dim}{name of dimension to be removed (default = "trans")}

\item{level}{level id from the dropped dimension that will be retained (default = "ind")}
}
\value{
a modified swissdata object with the specified dimension removed.
}
\description{
Drops an existing dimension from the provided Swissdata object.
}
\details{
Given an existing Swissdata object and a dimension name, this function
removes this dimension from the data by modifying the \code{x$data},
\code{x$meta$hierarchy}, \code{x$meta$labels}, and \code{x$meta$dim_order}
fields. When removing the dimesnion from \code{x$data} only the rows
corresponding to provided level argument are retained
}
\seealso{
\code{dim_add}
}
\author{
Christoph Sax
}
