% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_swissdata_s3.R
\name{read_swissdata_s3}
\alias{read_swissdata_s3}
\title{Download and Read swissdata set}
\usage{
read_swissdata_s3(set_id, bucket = "swissdata")
}
\description{
Download and Read swissdata set
}
\examples{
# works without autentication!
x <- sdtools::read_swissdata_s3("ch.fso.bapau")
}
