% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_swissdata.R
\name{test_swissdata}
\alias{test_swissdata}
\title{Test Swissdata}
\usage{
test_swissdata(x)
}
\arguments{
\item{x}{swissdata object}
}
\value{
If all tests passed the original swissdata object is returned invisibly.
Otherwise the execution is halted and an informative error message is displayed instead.
}
\description{
Extended set of tests for swissdata objects.
}
\details{
The function runs several tests on a swissdata object:\cr\cr
\enumerate{
\item "data" slot is a \code{tibble}\cr
\item "meta" slot is a \code{list}\cr
\item "meta" slot has all the required entries\cr
\item "meta" slot has no unexpected entries\cr
\item there are no missing values in \code{x$data$value}\cr
\item there are no missing values in \code{x$data$date}\cr
\item dates are specified as the starts of period\cr
\item last elements of "data" slot are "date" and "value"\cr
\item "meta" and "data" has identical dimensions\cr
\item all dimensions have unique labels in "meta"\cr
\item all dimensions in data have labels in "meta"\cr
\item there are no duplicated entries for the same date\cr
\item each dimension has it's own label\cr
\item languages are specified correctly\cr
\item source URL has a "http" in it\cr
}
}
\examples{
z <- adecco
test_swissdata(z)

}
\author{
Christoph Sax
}
