% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_sd.R
\name{theme_sd}
\alias{theme_sd}
\title{Create a Custom ggplot2 Theme}
\usage{
theme_sd(
  base_size = 11,
  base_family = "Open Sans",
  title_family = "Open Sans",
  axis_line = "none",
  axis_tick = "xy",
  major_tick_length = 3,
  grid = "xy",
  base_line_size = base_size/22,
  base_rect_size = base_size/22
)
}
\arguments{
\item{base_size}{Numeric. Base font size for the theme. Default is 11.}

\item{base_family}{Character. Font family for base text. Default is "Open Sans".}

\item{title_family}{Character. Font family for plot titles. Default is "Open Sans".}

\item{axis_line}{Character. Where to draw axis lines. Options: "xy", "x", "y", "none". Default is "none".}

\item{axis_tick}{Character. Where to draw axis ticks. Options: "xy", "x", "y", "none". Default is "xy".}

\item{major_tick_length}{Numeric. Major tick length in mm. Minor ticks length is half. Default is 3 mm.
To display ticks inside the plot use a negative number.}

\item{grid}{Character. Where to draw grid lines. Options: "xy", "x", "y", "none". Default is "xy".}

\item{base_line_size}{Numeric. Line size for base elements. Default is \code{base_size / 22}.}

\item{base_rect_size}{Numeric. Rectangle size for base elements. Default is \code{base_size / 22}.}
}
\value{
A ggplot2 theme object.
}
\description{
Creates a customized ggplot2 theme based on theme_minimal with enhanced typography
and customized grid lines. This theme uses the Roboto family for base text and
Roboto Slab for titles by default.
}
\details{
The theme checks for font availability and will fall back to system defaults if
required fonts are not installed. It will provide instructions for installing
missing fonts from Google Fonts.
}
\examples{
\dontrun{
library(ggplot2)

# Basic usage with defaults
ggplot(mtcars, aes(wt, mpg)) +
  geom_point() +
  theme_sd()

# Modify grid lines to show only vertical grid
ggplot(mtcars, aes(wt, mpg)) +
  geom_point() +
  theme_sd(grid = "y", axis_tick = "y", axis_line = "y")

# To display minor ticks use guides()
ggplot(mpg, aes(displ, hwy)) +
  geom_point() +
  guides(
    x = guide_axis(minor.ticks = TRUE),
    y = guide_axis(minor.ticks = TRUE)
  ) +
  scale_y_continuous(expand = expansion(0), minor_breaks = scales::breaks_width(1)) +
  scale_x_continuous(expand = expansion(0), minor_breaks = scales::breaks_width(0.2)) +
  theme_sd(grid = "xy", major_tick_length = -3)
}


}
