% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dim_add.R
\name{dim_add}
\alias{dim_add}
\title{Add New Dimension}
\usage{
dim_add(x, dim = "geo", level = "0", label = level)
}
\arguments{
\item{x}{swissdata object}

\item{dim}{name of dimension to be added (default = "geo")}

\item{level}{level id used for the newly added dimension (default = "0")}

\item{label}{English label for the newly added dimension (default = \code{level})}
}
\value{
a modified swissdata object that includes newly added dimension
}
\description{
Adds an additional dimension to the provided Swissdata object.
}
\details{
Given an existing Swissdata object and additional parameters of
dimension name, dimension level, and dimension label this function
incorporates the newly specified dimension by altering the \code{x$data},
\code{x$meta$hierarchy}, \code{x$meta$labels}, and \code{x$meta$dim_order}
fields.
}
\seealso{
\code{dim_drop}
}
\author{
Christoph Sax
}
