% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dim_add.R
\name{dim_add}
\alias{dim_add}
\alias{dim_drop}
\title{Add or Drop Dimensions}
\usage{
dim_add(x, dim = "geo", level = "0", label = level)

dim_drop(x, dim = "trans", level = "ind")
}
\arguments{
\item{x}{swissdata object}

\item{dim}{name of dimension to be added or removed}

\item{level}{single level ID to be used in the new dimension (\code{dim_add}), or
single level ID for which the data is filtered (\code{dim_drop}).}

\item{label}{English label for the newly added dimension}
}
\value{
a modified swissdata object
}
\description{
\code{dim_add} adds an additional dimension to a Swissdata object, \code{dim_drop}
removes a dimension.
}
\examples{
  # add new dimension "new_dim"
  z <- adecco
  z <- dim_add(z, dim = "new_dim", level = "nd", label = "newly added dim")
  z$data
  z$meta$dim_order
  z$meta$labels$new_dim
  z$meta$hierarchy$new_dim

  # drop existing dimension idx_type (collapse to level "sch")
  z <- dim_drop(z, dim = "idx_type", level = "sch")
  z$data
  z$meta$dim_order
  z$meta$labels$idx_type
  z$meta$hierarchy$idx_type

}
\author{
Christoph Sax
}
