% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marry.R
\name{marry}
\alias{marry}
\title{Marry Two Swissdata Objects}
\usage{
marry(x, y)
}
\arguments{
\item{x}{original swissdata object}

\item{y}{additional swissdata object}
}
\value{
a combined swissdata object
}
\description{
Combine the dimension levels from two swissdata objects.
}
\details{
The function adds dimension levels that are present only in \code{y} to \code{x}.
As a result in the case where the same dimensions or levels are present
in both of the objects - the one from \code{x} is selected.

\code{test_swissdata()} function is executed before returning the output
in order to make sure the combination of the two objects was successful.
}
\examples{
z  <- adecco
z2 <- level_rename(z, "idx_type", "ins", "new")
marry(z, z2)

}
\author{
Christoph Sax
}
