% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.swissdata}
\alias{print.swissdata}
\title{Print Swissdata Object}
\usage{
\method{print}{swissdata}(x, language = "en")
}
\arguments{
\item{x}{swissdata object}

\item{language}{language to be used for printing the labels}
}
\value{
invisibly returns the swissdata object
}
\description{
Function for printing swissdata objects
}
\examples{
z <- adecco
print(z)
print(z, languge = "de")

}
\seealso{
\code{str.swissdata}
}
\author{
Karolis KonceviÄius
}
