% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read}
\alias{read}
\alias{dataset_read_yaml}
\alias{dataset_read_json}
\alias{dataset_read_s3}
\title{Download and Read Swissdata Objects}
\usage{
dataset_read_yaml(set_path, test = TRUE)

dataset_read_json(set_path, test = TRUE)

dataset_read_s3(set_id, bucket = "swissdata", test = TRUE)
}
\arguments{
\item{set_path}{directory containing the files to be read}

\item{test}{should \code{dataset_validate()} be run on the newly-read object (default = TRUE)}

\item{set_id}{swissdata series id used for reading the object from "Amazon s3".}

\item{bucket}{name for "Amazon s3" (default = "swissdata")}
}
\value{
An object of class swissdata
}
\description{
Functions for obtaining swissdata objects from different sources.
Possible input formats include: yaml files, json files,
and Amazon Simple Storage Service (s3)
}
\examples{
  # read from yaml
  # set_path <- "some/path"
  # x <- dataset_read_yaml(set_path)

  # read from json
  # set_path <- "some/path"
  # x <- dataset_read_json(set_path)

  # read from s3
  # x <- dataset_read_s3("ch.fso.bapau")

}
\author{
Christoph Sax
}
