% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors_2.R
\name{sd_pal_c}
\alias{sd_pal_c}
\title{Continuous sd palette function}
\usage{
sd_pal_c(palette = "default", reverse = FALSE, ...)
}
\arguments{
\item{palette}{Character string, one of \code{"default"}, \code{"blues"}, or \code{"grays"}.}

\item{reverse}{Logical; if \code{TRUE}, the palette is reversed.}

\item{...}{Additional arguments passed to \code{grDevices::colorRampPalette()}.}
}
\value{
A function from \code{grDevices::colorRampPalette()} that generates
color vectors.
}
\description{
Returns a colorRampPalette based on an sd palette for continuous scales.
}
\examples{
pal <- sd_pal_c("blues")
pal(10)
}
