% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str.swissdata.R
\name{str.swissdata}
\alias{str.swissdata}
\title{Compact Display of the Swissdata Object}
\usage{
\method{str}{swissdata}(x, language = "en")
}
\arguments{
\item{x}{swissdata object}

\item{language}{language of the labels to be used for displaying the structure}
}
\value{
Does not return anything but displays the swissdata structure.
}
\description{
Compactly displays the internal hierarchy and labels of swissdata objects.
}
\examples{
z <- adecco
str(z)
str(z, languge = "de")

}
\author{
Karolis KonceviÄius
}
