% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.R
\name{write}
\alias{write}
\alias{write_swissdata}
\title{Write Swissdata Objects}
\usage{
write_swissdata(x, path_out, test = TRUE)
}
\arguments{
\item{x}{swissdata object to be written}

\item{path_out}{path for top-level directory for storing swissdata objects}

\item{test}{should \code{test_swissdata()} be run before writing the object (default = TRUE)}
}
\value{
If operation is successful the destination directory for the object is returned invisibly.
}
\description{
Functions for exporting swissdata objects by writing their contents to files.
}
\details{
In this function the argument \code{path_out} specifies the top level directory.
The directory containing all the files for the saved swissdata object will be stored within \code{path_out} (see examples).
}
\examples{
 z <- adecco
 root_dir <- tempdir()
 data_dir <- write_swissdata(z, dir, test = FALSE)
 x <- read_swissdata_yaml(dir)
 str(x)
 str(z)
 # remove all created directories
 unlink(root_dir, recursive = TRUE)

}
\author{
Christoph Sax
}
